Moving common definitions to shared header
