+++ miniapps/plasma/diffusion/makefile
@@ -38,6 +38,8 @@ COMMON_LIB += $(if $(MFEM_SHARED:YES=),,\
    $(if $(MFEM_USE_CUDA:YES=),$(CXX_XLINKER),$(CUDA_XLINKER))-rpath,$(abspath\
    $(MFEM_BUILD_DIR)/miniapps/common))
 
+COMMON_HDR = umansky.hpp
+
 COMMON_O=
 
 # Remove built-in rules
@@ -47,13 +49,14 @@ COMMON_O=
 all: $(MINIAPPS)
 
 # Rules for building the miniapps
-%: $(SRC)%.cpp $(COMMON_O) $(MFEM_LIB_FILE) $(CONFIG_MK) | lib-common
+%: $(SRC)%.cpp $(COMMON_HDR) $(COMMON_O) $(MFEM_LIB_FILE) $(CONFIG_MK) \
+	| lib-common
 	$(MFEM_CXX) $(MFEM_LINK_FLAGS) $< -o $@ $(COMMON_O) $(COMMON_LIB) \
         $(MFEM_LIBS)
 
 # Rules for compiling miniapp dependencies
 $(COMMON_O) $(addsuffix _solver.o,$(MINIAPPS)): \
-%.o: $(SRC)%.cpp $(SRC)%.hpp $(CONFIG_MK)
+%.o: $(SRC)%.cpp $(SRC)%.hpp $(COMMON_HDR) $(CONFIG_MK)
 	$(MFEM_CXX) $(MFEM_FLAGS) -c $(<) -o $(@)
 
 # Rule for building lib-common
